결제 정보 redis저장 ,결제 시간 제한 / 코드리팩토링

+++ order-service/src/main/java/com/example/orderservice/dto/WishListItemDto.java
@@ -11,5 +11,5 @@ public class WishListItemDto {
 
     private Long productId;
 
-    private Long count;
+    private int count;
 }
